Create structure.c
I'll Write a C program that defines a structure employee containing the details such as empno, empname, department name and salary. The structure has to store 20 employees in an organization. Use the appropriate method to define the above details and define a function that will display the contents?

Your Output should look similar to this

Name                Age       Phone Number     Salary

Chirag                24         1234567788       20000

Arnav                  31        1234567891        56000


Shivam               45         8881101111     30500

       That's all my cap bro 😅